avoid overflow

+++ crypto/math/src/gpu/metal/fft.rs
@@ -736,7 +736,7 @@ mod tests {
     fn test_fft_order_validation() {
         // Order 32 would make n = 2^32, which truncates to 0 when cast to u32
         // for Metal shader parameters. Verify the truncation is real:
-        let n: usize = 1 << 32u64;
+        let n: u64 = 1u64 << 32;
         assert_eq!(
             n as u32, 0,
             "2^32 as u32 must be 0 — this is the bug we're preventing"
